ServiceAccounts: Fix token-apikey cross deletion

+++ pkg/services/sqlstore/apikey.go
@@ -61,7 +61,7 @@ func (ss *SQLStore) DeleteApiKey(ctx context.Context, cmd *models.DeleteApiKeyCo
 }
 
 func deleteAPIKey(sess *DBSession, id, orgID int64) error {
-	rawSQL := "DELETE FROM api_key WHERE id=? and org_id=?"
+	rawSQL := "DELETE FROM api_key WHERE id=? and org_id=? and service_account_id is null"
 	result, err := sess.Exec(rawSQL, id, orgID)
 	if err != nil {
 		return err
